requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database dependencies
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
websockets==11.0.3

# Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# AI and ML dependencies
openai>=1.3.0
transformers>=4.30.0
torch>=2.0.0
numpy>=1.21.0
scikit-learn>=1.2.0

# LangChain and LangGraph for agentic AI
langchain>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.20
typing-extensions>=4.8.0

# Text processing
textblob==0.17.1
vaderSentiment==3.3.2

# Data processing
pandas>=2.0.0

# HTTP client
httpx>=0.25.0

# Audio processing for voice journaling
SpeechRecognition>=3.10.0
pydub>=0.25.1
pyaudio>=0.2.11

# Testing (optional for production)
pytest>=7.4.0
pytest-asyncio>=0.21.0
